Bump SwiftLint, Mockolo and LicensePlist
