--- Tools/UhooiPicBookMockolo/Package.resolved
@@ -1,43 +1,0 @@
-{
-  "object": {
-    "pins": [
-      {
-        "package": "Mockolo",
-        "repositoryURL": "https://github.com/uber/mockolo",
-        "state": {
-          "branch": null,
-          "revision": "e1d3e7b34ec44239d3eb022d109fd3dc9d6ade40",
-          "version": "1.6.1"
-        }
-      },
-      {
-        "package": "swift-argument-parser",
-        "repositoryURL": "https://github.com/apple/swift-argument-parser",
-        "state": {
-          "branch": null,
-          "revision": "d2930e8fcf9c33162b9fcc1d522bc975e2d4179b",
-          "version": "1.0.1"
-        }
-      },
-      {
-        "package": "SwiftSyntax",
-        "repositoryURL": "https://github.com/apple/swift-syntax.git",
-        "state": {
-          "branch": null,
-          "revision": "75e60475d9d8fd5bbc16a12e0eaa2cb01b0c322e",
-          "version": "0.50500.0"
-        }
-      },
-      {
-        "package": "swift-tools-support-core",
-        "repositoryURL": "https://github.com/apple/swift-tools-support-core.git",
-        "state": {
-          "branch": null,
-          "revision": "3b6b97d612b56e25d80d0807f5bc38ea08b7bdf3",
-          "version": "0.2.3"
-        }
-      }
-    ]
-  },
-  "version": 1
-}
