+++ Tools/UhooiPicBookTools/Package.swift
@@ -8,13 +8,13 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/yonaskolb/XcodeGen", .exact("2.25.0")),
-        .package(url: "https://github.com/realm/SwiftLint", .exact("0.44.0")),
+        .package(url: "https://github.com/realm/SwiftLint", .exact("0.45.0")),
         .package(url: "https://github.com/IBDecodable/IBLinter", .exact("0.4.27")),
         .package(url: "https://github.com/fromkk/SpellChecker", .exact("0.1.0")),
-        // .package(url: "https://github.com/uber/mockolo", .exact("1.6.1")),
-        .package(url: "https://github.com/mono0926/LicensePlist", .exact("3.13.0")),
+        .package(url: "https://github.com/uber/mockolo", .exact("1.6.2")),
+        .package(url: "https://github.com/mono0926/LicensePlist", .exact("3.14.4")),
         .package(url: "https://github.com/mac-cain13/R.swift", .exact("5.4.0")),
-        .package(url: "https://github.com/thii/xcbeautify", .exact("0.9.1"))
+        // .package(url: "https://github.com/thii/xcbeautify", .exact("0.9.1"))
     ],
     targets: [.target(name: "UhooiPicBookTools", path: "")]
 )
